requirements.txt
beautifulsoup4==4.10.0
Bottleneck==1.3.2
certifi==2021.10.8
charset-normalizer==2.0.7
gevent==21.8.0
greenlet==1.1.2
grequests==0.6.0
idna==3.3
mkl-fft==1.3.1
mkl-service==2.4.0
pandas==1.3.4
PyMySQL==1.0.2
python-dotenv==0.19.2
pytz==2021.3
regex==2021.10.23
requests==2.26.0
soupsieve==2.2.1
urllib3==1.26.7
zope.event==4.5.0
zope.interface==5.4.0
